+++ streamlit_app.py
@@ -4,7 +4,7 @@ from typing import Optional
 import altair as alt
 import pandas as pd
 import streamlit as st
-from st_aggrid import AgGrid, GridOptionsBuilder
+from st_aggrid import AgGrid, DataReturnMode, GridOptionsBuilder
 from st_aggrid.shared import GridUpdateMode
 
 from refida import data as dm
@@ -23,22 +23,26 @@ def data_section():
 
     data = dm.get_etl_data()
     if data is not None:
-        selection = show_data(data)
+        grid = show_data_grid(data)
 
-        if selection and len(selection["selected_rows"]) > 0:
-            doc = selection["selected_rows"][0]
-            show_doc(doc)
+        selection = grid["selected_rows"]
+
+        if selection:
+            show_data(pd.DataFrame(selection))
+        else:
+            show_data(grid["data"])
 
 
-def show_data(data: pd.DataFrame) -> AgGrid:
+def show_data_grid(data: pd.DataFrame) -> dict:
+    data = data.drop(columns=["names", "compressed"])
     options = GridOptionsBuilder.from_dataframe(
         data, enableRowGroup=True, enableValue=True
     )
-    # options.configure_side_bar()
-    options.configure_selection("single")
+    options.configure_selection("multiple", use_checkbox=True)
 
-    selection = AgGrid(
+    grid = AgGrid(
         data,
+        data_return_mode=DataReturnMode.FILTERED,
         enable_enterprise_modules=True,
         gridOptions=options.build(),
         height=300,
@@ -47,44 +51,57 @@ def show_data(data: pd.DataFrame) -> AgGrid:
         allow_unsafe_jscode=False,
     )
 
-    return selection
+    return grid
 
 
-def show_doc(doc: pd.DataFrame):
-    st.header(doc["title"])
+def show_data(data: pd.DataFrame):
+    n_rows = data.shape[0]
+    topic_score = "score"
 
-    st.subheader("Summary")
-    summary = get_summary(doc["id"])
-    if summary is not None:
-        st.write(summary)
+    if n_rows == 0:
+        st.warning("No data found")
+        return
+
+    st.header("About the data")
+
+    if n_rows > 1:
+        st.info("Multiple documents available, showing aggregate information")
+        topic_score = "mean(score)"
     else:
-        st.write(doc["summary"])
+        doc = data.iloc[0]
+        st.header(doc["title"])
+        st.subheader("About the case study")
+        summary = get_summary(tuple([doc["id"]]))
+        if summary is not None:
+            st.write(summary)
+        else:
+            st.write(doc["summary"])
 
-    topics = get_topics(doc["id"])
+    topics = get_topics(tuple(data["id"].values.tolist()))
     if topics is not None:
-        st.subheader("Topics")
+        st.subheader("Impact categories")
         st.altair_chart(
             alt.Chart(topics)
             .mark_bar(tooltip=True)
-            .encode(x="score", y="topic", color="topic"),
+            .encode(x=topic_score, y="topic", color="topic"),
             use_container_width=True,
         )
 
 
 @lru_cache(maxsize=256)
-def get_summary(id: str) -> Optional[str]:
+def get_summary(ids: tuple[str]) -> Optional[str]:
     data = dm.get_summaries_data()
 
     if data is not None:
-        summary = get_rows_by_id(data, id)
+        summary = get_rows_by_id(data, ids)
         if summary is not None:
             return summary["summary"].iloc[0]
 
     return None
 
 
-def get_rows_by_id(data: pd.DataFrame, id: str) -> Optional[pd.DataFrame]:
-    rows = data[data["id"] == id]
+def get_rows_by_id(data: pd.DataFrame, ids: tuple[str]) -> Optional[pd.DataFrame]:
+    rows = data[data["id"].isin(ids)]
 
     if rows is not None:
         return rows
@@ -93,11 +110,11 @@ def get_rows_by_id(data: pd.DataFrame, id: str) -> Optional[pd.DataFrame]:
 
 
 @lru_cache(maxsize=256)
-def get_topics(id: str) -> Optional[pd.DataFrame]:
+def get_topics(ids: tuple[str]) -> Optional[pd.DataFrame]:
     data = dm.get_topics_data()
 
     if data is not None:
-        topics = get_rows_by_id(data, id)
+        topics = get_rows_by_id(data, ids)
         if topics is not None:
             return topics[["topic", "score"]]