+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Command and module to extract, transform and load data.
 - Text summarisation.
 - Topic classification.
+- Dashboard with data, summary and topics chart.